backend/requirements.txt
# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Security & Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=44.0.0

# Utilities
python-dotenv>=1.0.1
python-multipart>=0.0.20
aiofiles>=24.1.0

# Testing (optional for development)
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
